internal/contour remove contour.Translator
Ingress class variable has moved to contour.DAGAdapter, still not hooked
up.
